BTT-68-Login-POST-endpoint

+++ backend/src/main/java/com/rkrs/bikethefttracker/security/AuthenticationService.java
@@ -69,17 +69,26 @@ public class AuthenticationService {
         return new LoginResult(loginResponse, accessToken, refreshToken);
     }
 
-    public void logout() {
-        String jtiString = jwtContext.getClaims().getId();
+    public void logout(String refreshToken) {
+        if (refreshToken == null) {
+            return;
+        }
+
+        Claims claims = jwtService.parseToken(refreshToken);
+        String jtiString = claims.getId();
         if (jtiString == null) {
-            throw new InvalidRefreshTokenException("Invalid refresh token.");
+            return;
         }
 
         refreshTokenService.deleteByJwtId(getJti(jtiString));
     }
 
-    public JwtToken renewAccessToken() {
-        Claims claims = jwtContext.getClaims();
+    public JwtToken renewAccessToken(String refreshTokenCookie) {
+        if (refreshTokenCookie == null) {
+            throw new InvalidRefreshTokenException("Invalid refresh token. Please log in again.");
+        }
+
+        Claims claims = jwtService.parseToken(refreshTokenCookie);
         RefreshToken refreshToken = this.getValidRefreshToken(claims);
 
         User user = refreshToken.getUser();
